back_end/requirements.txt
# Core Python utilities
numpy==1.26.4
pandas==2.3.1
matplotlib==3.10.5
scipy==1.16.1
statsmodels==0.14.5
typing-extensions==4.14.1
annotated-types==0.7.0
typing-inspection==0.4.1

# Web & API
fastapi==0.116.1
uvicorn==0.35.0
h11==0.16.0
click==8.2.1
starlette==0.47.2
anyio==4.10.0
requests==2.32.5
beautifulsoup4==4.13.4
PyPDF2==3.0.1
reportlab==4.4.3
python-dotenv==1.0.1
python-multipart==0.0.20

# Redis / Celery
redis==6.4.0
celery==5.5.3
billiard==4.2.1
kombu==5.5.4
vine==5.1.0

# LLM / LangChain stack
langchain==0.2.6
langchain-groq==0.1.10
langchain-openai==0.1.8
langchain-community==0.2.6
langchain-core==0.2.43
langchain-cohere==0.2.4
langchain-experimental==0.0.63
langchain-text-splitters==0.2.2
langsmith==0.1.147

# Embeddings / Vector stores
faiss-cpu==1.9.0
sentence-transformers==5.1.0
rank-bm25==0.2.2
huggingface-hub==0.34.4
transformers==4.55.2
accelerate==1.28.0
torch==2.8.0+cpu
scikit-learn==1.7.1

# Experimentation / Stats
sympy==1.13.3
tenacity==8.5.0
tabulate==0.9.0

# Optional / Extras
alembic==1.16.4
SQLAlchemy==2.0.43
Jinja2==3.1.4
Mako==1.3.10
networkx==3.3
regex==2025.7.34
tiktoken==0.11.0
cohere==5.17.0
openai==1.100.2
orjson==3.11.2
ujson==5.11.0
watchfiles==1.1.0
websockets==15.0.1

# CLI / Tooling
typer==0.16.1
prompt_toolkit==3.0.51
wcwidth==0.2.13
rich==14.1.0
rich-toolkit==0.15.0

# Utilities
filelock==3.13.1
joblib==1.5.1
tqdm==4.67.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# HTTP / Networking
httpx==0.28.1
httpcore==1.0.9
httptools==0.6.4
httpx-sse==0.4.0
aiohttp==3.12.15
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
multidict==6.6.4
yarl==1.20.1
idna==3.10
urllib3==2.5.0
charset-normalizer==3.4.3
certifi==2025.8.3
dnspython==2.7.0

# Validation / Parsing
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
marshmallow==3.26.1
attrs==25.3.0
email_validator==2.2.0
PyYAML==6.0.2

# Parsing / Markup
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
soupsieve==2.7

# Plotting / Math
contourpy==1.3.3
cycler==0.12.1
kiwisolver==1.4.9
pillow==11.3.0
fonttools==4.59.1
mpmath==1.3.0

# Testing
pytest==8.4.1
iniconfig==2.1.0
pluggy==1.6.0

# Async / Performance
uvloop==0.21.0
structlog==25.4.0
propcache==0.3.2
greenlet==3.2.4
